Update notifier discussion
Refresh the discussion of data made available through the 1.5.x history when addressing observers and notifiers.
Add discussion of camelCase methods and having access to modify data within malleable parameters.
Provide a more accurate explanation about attaching to notifiers in the load sequence.  Previously it seemed that it was identified that the code having the notifier had to be "loaded" before attempting to observe it; however, this has led to improper initial implementation of monitoring notifiers that are within class construction methods.  The important part is to have the observer in place before the notifier is executed not after.
Revised the real-world example to account for additional notifiers that are internal to the shopping_cart class that would remove a product without executing the remove method.  Also revised to allow the free product presence to fully toggle with the threshold, although this revision does not necessarily respect keeping the free product out of the cart in the following sequence: threshold met, product added, customer removes free product, cart modified to go less than the threshold, then the cart quantity increased to pass the threshold but keep the free product out.  By increasing the cart quantity above the threshold again, even though the customer previously removed the free product, it would be added back in.
Offered some code writing suggestions and ideas.